# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Studify API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupId}/invite:
    post:
      operationId: sendInvite
      tags: [invites]
      summary: Send an invite to a user by email (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInviteBody"
      responses:
        "201":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /invites/{inviteId}/accept:
    post:
      operationId: acceptInvite
      tags: [invites]
      summary: Accept a group invite
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invites/{inviteId}/reject:
    post:
      operationId: rejectInvite
      tags: [invites]
      summary: Reject a group invite
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adminId:
          type: integer
        adminName:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminId
        - adminName
        - memberCount
        - createdAt
...
    SendInviteBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    Invite:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        groupName:
          type: string
        invitedById:
          type: integer
        invitedByName:
          type: string
        invitedUserId:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - groupName
        - invitedById
        - invitedByName
        - invitedUserId
        - status
        - createdAt
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        streak:
          type: integer
        totalHours:
          type: number
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - streak
        - totalHours
        - groups
        - createdAt