# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Invite {
  id: number;
  groupId: number;
  groupName: string;
  invitedById: number;
  invitedByName: string;
  invitedUserId: number;
  status: string;
  createdAt: Date;
}