# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Group } from "./group";

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  streak: number;
  totalHours: number;
  groups: Group[];
  createdAt: Date;
}